backend/requirements.txt
firebase-admin
accelerate==0.18.0
appdirs==1.4.4
asgiref==3.7.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
clean-fid==0.1.35
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
diffusers==0.14.0
Django==5.0.1
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
docker-pycreds==0.4.0
drf-yasg==1.21.7
filelock==3.13.1
fonttools==4.47.2
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.41
google-api-core==2.15.0
google-api-python-client==2.115.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
huggingface-hub==0.20.3
idna==3.6
importlib-metadata==7.0.1
# inflection==0.5.1
# Jinja2==3.1.3
# kiwisolver==1.4.5
# MarkupSafe==2.1.4
# # matplotlib==3.7.1
# mpmath==1.3.0
# networkx==3.2.1
# ninja==1.11.1.1
# numpy==1.26.3
# opencv-python==4.7.0.72
# packaging==23.2
# pathtools==0.1.2
# pillow==10.2.0
# protobuf==4.25.2
# psutil==5.9.8
# pyasn1==0.5.1
# pyasn1-modules==0.3.0
# pyparsing==3.1.1
# python-dateutil==2.8.2
# pytz==2023.3.post1
# PyYAML==6.0.1
# regex==2023.12.25
requests==2.31.0
# rsa==4.9
# scipy==1.12.0
# sentry-sdk==1.39.2
# setproctitle==1.3.3
# six==1.16.0
# smmap==5.0.1
# sqlparse==0.4.4
# sympy==1.12
# tokenizers==0.13.3
torch==2.0.1
# torchmetrics==0.11.4
torchvision==0.15.2
# # tqdm==4.66.1
transformers==4.27.3
# typing_extensions==4.9.0
# tzdata==2023.4
# uritemplate==4.1.1
# urllib3==2.1.0
# wandb==0.14.0
# xformers==0.0.22
# zipp==3.17.0
